[Downloader] Only attempt to copy the download if theres a source location
Closes https://github.com/CocoaPods/CocoaPods/issues/3433

+++ lib/cocoapods/downloader.rb
@@ -32,7 +32,7 @@ module Pod
       cache_path, tmp_cache = Pathname(Dir.mktmpdir), true unless cache_path
       cache = Cache.new(cache_path)
       result = cache.download_pod(request)
-      if target
+      if target && result.location
         UI.message "Copying #{request.name} from `#{result.location}` to #{UI.path target}", '> ' do
           FileUtils.rm_rf target
           FileUtils.cp_r(result.location, target)
